requirements.txt
streamlit>=1.24.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
langchain>=0.0.267
PyPDF2>=3.0.0
chromadb
faiss-cpu>=1.7.4
langchain_google_genai
langchain-community
